deployment/requirements.txt
azure-functions==1.17.0
azure-functions-worker==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2 